+++ src/applications/vaos/tests/components/VAOSApp/AppUnavailable.unit.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import AppUnavailable from '../../components/AppUnavailable';
+import AppUnavailable from '../../../components/VAOSApp/AppUnavailable';
 
 describe('VAOS <AppUnavailable>', () => {
   it('should render', () => {
